Spring security project  使用@Secured 注解限制方法的调用

+++ 006Spring/SpringSecuritySpittr/src/main/java/con/hef/spittr/service/impl/SpitterServiceimpl.java
@@ -4,22 +4,37 @@ import con.hef.spittr.dao.SpitterDao;
 import con.hef.spittr.domain.Spitter;
 import con.hef.spittr.service.SpitterService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.annotation.Secured;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.persistence.EntityNotFoundException;
+import java.util.logging.Logger;
+
 @Component
-@Transactional
 public class SpitterServiceimpl implements SpitterService {
+
+    private static final Logger logger = Logger.getLogger("SpitterServiceimpl");
 
     @Autowired
     private SpitterDao spitterDao;
 
     @Override
+    @Transactional
     public Spitter findSpitterById(Long id) {
-        return spitterDao.getOne(id);
+        try {
+            Spitter spitter = spitterDao.getOne(id);;
+            return spitter;
+        }catch (Exception e){
+            logger.info("entity not found.");
+        }
+        return null;
     }
 
     @Override
+    @Transactional
+    @Secured(value = {"ROLE_SPITTER", "ROLE_ADMIN"})
     public Spitter saveSpitter(Spitter spitter) {
         Spitter save = spitterDao.save(spitter);
         return save;
@@ -28,6 +43,8 @@ public class SpitterServiceimpl implements SpitterService {
     @Override
     public void deleteSpitter(long id) {
         Spitter spitter = findSpitterById(id);
-        spitterDao.delete(spitter);
+        if (spitter!=null){
+            spitterDao.delete(spitter);
+        }
     }
 }
